requirements.txt
streamlit
pandas
numpy
fpdf2
plotly-express